+++ vnc_winsrc/vncviewer/vncviewer.cpp
@@ -1,3 +1,4 @@
+//  Copyright (C) 2010 D. R. Commander. All Rights Reserved.
 //  Copyright (C) 1999 AT&T Laboratories Cambridge. All Rights Reserved.
 //
 //  This file is part of the VNC system.
@@ -85,7 +86,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, PSTR szCmdLine,
 
 	vnclog.Print(3, _T("Exiting\n"));
 
-	return msg.wParam;
+	return (int)msg.wParam;
 }
